+++ packages/hot-updater/src/commands/init.ts
@@ -2,7 +2,7 @@ import { printBanner } from "@/components/banner";
 import { ensureInstallPackages } from "@/utils/ensureInstallPackages";
 import { isCancel, select } from "@clack/prompts";
 import { initCloudflareD1R2 } from "./init/cloudflareD1R2";
-import { initSupabase } from "./init/supabase";
+import { initSupabase } from "./init/initSupabase";
 
 const REQUIRED_PACKAGES = {
   dependencies: ["@hot-updater/react-native"],