+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 0.5.10 (2025-01-22)
+
+### 🩹 Fixes
+
+- **supabase:** improve error message for upload failures in supabaseStorage ([ab6f9f5](https://github.com/gronxb/hot-updater/commit/ab6f9f5))
+
+### ❤️ Thank You
+
+- gronxb
+
+## 0.5.9 (2025-01-22)
+
+### 🚀 Features
+
+- **cli:** get-plugin-env ([#70](https://github.com/gronxb/hot-updater/pull/70))
+- **metro:** add `entryFile`, `sourcemap` parameter to metro() ([#69](https://github.com/gronxb/hot-updater/pull/69))
+
+### ❤️ Thank You
+
+- Sungyu Kang
+
 ## 0.5.8 (2025-01-21)
 
 ### 🩹 Fixes